Rename: 랜딩,리스트 페이지 디렉토리 정리
